modified tests and end points
