+++ api/__init__.py
@@ -1,0 +1,3 @@
+from flask import Flask
+from api import views
+app = Flask(__name__)
