--- app/__init__.py
@@ -1,3 +1,0 @@
-from flask import Flask
-from app import views
-my_app = Flask(__name__)
