--- main.py
@@ -1,5 +1,0 @@
-from app.views import my_app
-
-if __name__ == "__main__":
-	my_app.run(debug = True)
-
